Move state information from helpers into state machine
Don't especially like that the display is there, but will work on moving
the display stuff back into a helper when the time comes. I expect this
struct to go away, as well, eventually moving to using the raw methods
from the state machine. We'll see how it comes out, though.